Add files via upload
QUESTION :
Write a python program to create a conversational bot 
DESCRIPTION :
The bot is a virtual assistant who gives you a tour around a bookstore and 
depending on your choice of genre it displays to you a number of books. You can 
add the book that you like to a cart. And at the end it tells you the total amount and 
displays the books that you have bought. 
LIMITATIONS :
In the while loop the customer can only go to a genre below the one that they are 
traversing. It cannot go to a genre that is mentioned above . 
FUTURE EXPANSION PROSPECT :
In the future , depending on the books that you like the assistant might be able to 
give similar recommendation with the addition of more genres. The assistant might 
also be able to give the customer a short description of the book that he/she likes 
and can show other books of the same author.